Move files to common dir
