+++ chunkymap-core/src/main/java/de/lemaik/chunkymap/dynmap/ChunkyMap.java
@@ -8,7 +8,6 @@ import de.lemaik.chunkymap.util.MinecraftDownloader;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
-import java.io.InputStream;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
