+++ chunkymap-core/src/main/java/de/lemaik/chunkymap/rendering/SilentTaskTracker.java
