+++ chunkymap-core/src/main/java/de/lemaik/chunkymap/util/MinecraftDownloader.java
